added Event to fk check(s) in Month, Week, & Day also added CelestialEvent to fk check(s) where I missed it

+++ src/main/java/com/mcommings/campaigner/services/calendar/DayService.java
@@ -3,6 +3,7 @@ package com.mcommings.campaigner.services.calendar;
 import com.mcommings.campaigner.interfaces.calendar.IDay;
 import com.mcommings.campaigner.models.RepositoryHelper;
 import com.mcommings.campaigner.models.calendar.Day;
+import com.mcommings.campaigner.repositories.IEventRepository;
 import com.mcommings.campaigner.repositories.calendar.ICelestialEventRepository;
 import com.mcommings.campaigner.repositories.calendar.IDayRepository;
 import com.mcommings.campaigner.repositories.calendar.IWeekRepository;
@@ -17,6 +18,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import static com.mcommings.campaigner.enums.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ForeignKey.FK_DAY;
 
 @Service
 public class DayService implements IDay {
@@ -24,13 +26,15 @@ public class DayService implements IDay {
     private final IDayRepository dayRepository;
     private final IWeekRepository weekRepository;
     private final ICelestialEventRepository celestialEventRepository;
+    private final IEventRepository eventRepository;
 
     @Autowired
     public DayService(IDayRepository dayRepository, IWeekRepository weekRepository,
-                      ICelestialEventRepository celestialEventRepository) {
+                      ICelestialEventRepository celestialEventRepository, IEventRepository eventRepository) {
         this.dayRepository = dayRepository;
         this.weekRepository = weekRepository;
         this.celestialEventRepository = celestialEventRepository;
+        this.eventRepository = eventRepository;
     }
 
     @Override
@@ -57,10 +61,9 @@ public class DayService implements IDay {
         if (RepositoryHelper.cannotFindId(dayRepository, dayId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-//        TODO: uncomment once Event has been added
-//        if (RepositoryHelper.isForeignKey(getReposWhereDayIsAForeignKey(), FK_DAY.columnName, dayId)) {
-//            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
-//        }
+        if (RepositoryHelper.isForeignKey(getReposWhereDayIsAForeignKey(), FK_DAY.columnName, dayId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
         dayRepository.deleteById(dayId);
     }
 
@@ -79,11 +82,10 @@ public class DayService implements IDay {
         dayToUpdate.setFk_week(day.getFk_week());
     }
 
-//    TODO: uncomment once Event has been added
-//    private List<CrudRepository> getReposWhereDayIsAForeignKey() {
-//        List<CrudRepository> repositories = new ArrayList<>(Arrays.asList(celestialEventRepository, eventRepository));
-//        return repositories;
-//    }
+    private List<CrudRepository> getReposWhereDayIsAForeignKey() {
+        List<CrudRepository> repositories = new ArrayList<>(Arrays.asList(celestialEventRepository, eventRepository));
+        return repositories;
+    }
 
     private List<CrudRepository> getListOfForeignKeyRepositories() {
         List<CrudRepository> repositories = new ArrayList<>(Arrays.asList(weekRepository));
